If we can't parse our default personality, throw an exception and do not continue.

+++ SpeechResponder/Personality.cs
@@ -145,16 +145,23 @@ namespace EddiSpeechResponder
                 }
                 catch (Exception e)
                 {
-                    // malformed JSON for some reason: rename so that the user can examine and fix it.
-                    string newFileName = filename + ".malformed";
-                    if (File.Exists(newFileName))
+                    if (!isDefault)
                     {
-                        // no point keeping a history: only the latest is likely to be useful. Pro users will be using version control anyway.
-                        File.Delete(newFileName);
-                    }
-                    File.Move(filename, newFileName);
+                        // malformed JSON for some reason: rename so that the user can examine and fix it.
+                        string newFileName = filename + ".malformed";
+                        if (File.Exists(newFileName))
+                        {
+                            // no point keeping a history: only the latest is likely to be useful. Pro users will be using version control anyway.
+                            File.Delete(newFileName);
+                        }
+                        File.Move(filename, newFileName);
 
-                    Logging.Error($"Could not parse \"{filename}\": moved to \"{newFileName}\". Error was \"{e.Message}\"");
+                        Logging.Error($"Could not parse \"{filename}\": moved to \"{newFileName}\". Error was \"{e.Message}\"");
+                    }
+                    else
+                    {
+                        throw new FormatException("Could not parse default personality (eddi.json)");
+                    }
                 }
             }
